Factor down configureStoe() and root reducer
